+++ src/pat/inject/inject.js
@@ -476,26 +476,27 @@ define([
                         {selector: cfg.target});
                 return false;
             }
-            if (cfg.action === "content")
+            if (cfg.action === "content") {
                 $target.empty().append($source);
-            else if (cfg.action === "element")
+            } else if (cfg.action === "element") {
                 $target.replaceWith($source);
-            else
+            } else {
                 $target[method]($source);
-
+            }
             return true;
         },
 
         _sourcesFromHtml: function inject_sourcesFromHtml(html, url, sources) {
             var $html = inject._parseRawHtml(html, url);
             return sources.map(function inject_sourcesFromHtml_map(source) {
-                if (source === "body")
+                if (source === "body") {
                     source = "#__original_body";
-
+                }
                 var $source = $html.find(source);
 
-                if ($source.length === 0)
+                if ($source.length === 0) {
                     log.warn("No source elements for selector:", source, $html);
+                }
 
                 $source.find("a[href^=\"#\"]").each(function () {
                     var href = this.getAttribute("href");
@@ -507,12 +508,12 @@ define([
                     }
                     // Skip in-document links pointing to an id that is inside
                     // this fragment.
-                    if (href.length === 1)  // Special case for top-of-page links
+                    if (href.length === 1) { // Special case for top-of-page links
                         this.href=url;
-                    else if (!$source.find(href).length)
+                    } else if (!$source.find(href).length) {
                         this.href=url+href;
+                    }
                 });
-
                 return $source;
             });
         },
@@ -618,21 +619,19 @@ define([
                 log.error("Error rebasing urls", e);
             }
             var $html = $("<div/>").html(clean_html);
-
-            if ($html.children().length === 0)
+            if ($html.children().length === 0) {
                 log.warn("Parsing html resulted in empty jquery object:", clean_html);
-
+            }
             return $html;
         },
 
         // XXX: hack
         _initAutoloadVisible: function inject_initAutoloadVisible($el) {
-            // ignore executed autoloads
-            if ($el.data("pat-inject-autoloaded"))
+            if ($el.data("pat-inject-autoloaded")) {
+                // ignore executed autoloads
                 return false;
-
-            var $scrollable = $el.parents(":scrollable"),
-                checkVisibility;
+            }
+            var $scrollable = $el.parents(":scrollable"), checkVisibility;
 
             // function to trigger the autoload and mark as triggered
             function trigger() {
@@ -646,8 +645,9 @@ define([
                 // if scrollable parent and visible -> trigger it
                 // we only look at the closest scrollable parent, no nesting
                 checkVisibility = utils.debounce(function inject_checkVisibility_scrollable() {
-                    if ($el.data("patterns.autoload"))
+                    if ($el.data("patterns.autoload") || !$.contains(document, $el[0])) {
                         return false;
+                    }
                     var reltop = $el.offset().top - $scrollable.offset().top - 1000,
                         doTrigger = reltop <= $scrollable.innerHeight();
                     if (doTrigger) {
@@ -659,27 +659,28 @@ define([
                     }
                     return false;
                 }, 100);
-                if (checkVisibility())
+                if (checkVisibility()) {
                     return true;
-
+                }
                 // wait to become visible - again only immediate scrollable parent
                 $($scrollable[0]).on("scroll", checkVisibility);
                 $(window).on("resize.pat-autoload", checkVisibility);
             } else {
                 // Use case 2: scrolling the entire page
                 checkVisibility = utils.debounce(function inject_checkVisibility_not_scrollable() {
-                    if ($el.data("patterns.autoload"))
+                    if ($el.data("patterns.autoload")) {
                         return false;
-                    if (!utils.elementInViewport($el[0]))
+                    }
+                    if (!utils.elementInViewport($el[0])) {
                         return false;
-
+                    }
                     $(window).off(".pat-autoload", checkVisibility);
                     return trigger();
                 }, 100);
-                if (checkVisibility())
+                if (checkVisibility()) {
                     return true;
-                $(window).on("resize.pat-autoload scroll.pat-autoload",
-                        checkVisibility);
+                }
+                $(window).on("resize.pat-autoload scroll.pat-autoload", checkVisibility);
             }
             return false;
         },
@@ -691,7 +692,6 @@ define([
 
         callTypeHandler: function inject_callTypeHandler(type, fn, context, params) {
             type = type || "html";
-
             if (inject.handlers[type] && $.isFunction(inject.handlers[type][fn])) {
                 return inject.handlers[type][fn].apply(context, params);
             } else {
@@ -749,7 +749,6 @@ define([
             log.debug(e);
         }
     }
-
     registry.register(inject);
     return inject;
 });